273 - Update item association
- Cleanup the confirmation box modal
Removed couple of duplicate values since they are not relevant in the
confirmation box modal.

- Create sub directories and components for all required files

Created itemActions and itemContent files for DetailsCard.jsx component
Created components to flush out item association
Extracted CategoryItem and MaintenancePlanItem
Extracted category item component into its own sub children components
Extracted maintenace plan item component into its own sub children components

- Add pluralize function.
The function pluralize will add 's' character to the end of the string if
the size of the object to be considered is > 1.

- Fix lint issues and tech debt

Fixed lint issues for item-details page
Fixed tech debt and file naming conventions
Fixed table issue where the table was not overflowing to the right by
wrapping the container with Paper component and providing overflow
Fixed lint issues across the app especially with the newly built function
pluralize.
Fixed folder structure
Cleanup table component for action buttons
Flush table component into its own directives
