+++ frontend/src/Store.js
@@ -8,7 +8,7 @@ import notes from './features/Notes/notesSlice';
 import categories from './features/Categories/categoriesSlice';
 import categoryItemDetails from './features/CategoryItemDetails/categoryItemDetailsSlice';
 import maintenance from './features/MaintenancePlanList/maintenanceSlice';
-import maintenancePlanItem from './features/MaintenancePlanDetails/maintenancePlanItemSlice';
+import maintenancePlanItem from './features/MaintenancePlanItemDetails/maintenancePlanItemSlice';
 import inventory from './features/InventoryList/inventorySlice';
 import summary from './features/Home/SummarySlice';
 import reports from './features/Reports/reportSlice';
@@ -19,7 +19,7 @@ import notesSaga from './features/Notes/notesSaga';
 import categoriesSaga from './features/Categories/categoriesSaga';
 import categoryItemSaga from './features/CategoryItemDetails/categoryItemDetailsSaga';
 import maintenanceSaga from './features/MaintenancePlanList/maintenanceSaga';
-import maintenancePlanItemSaga from './features/MaintenancePlanDetails/maintenancePlanItemSaga';
+import maintenancePlanItemSaga from './features/MaintenancePlanItemDetails/maintenancePlanItemSaga';
 import inventorySaga from './features/InventoryList/inventorySaga';
 import summarySaga from './features/Home/SummarySaga';
 import reportsSaga from './features/Reports/reportSaga';
