+++ frontend/src/features/CategoryItemDetails/CategoryItemDetailsContent/CategoryItemDetailsGraph.jsx
@@ -1,16 +1,17 @@
-import { Box } from '@mui/material';
-import BarChart from '../../../../common/Chart/BarChart';
-import RowHeader from '../../../../common/RowHeader';
+import { Box, Paper } from '@mui/material';
 
-export default function ItemGraph({ totalItems }) {
+import RowHeader from '../../../common/RowHeader';
+import BarChart from '../../../common/Chart/BarChart';
+
+export default function CategoryItemDetailsGraph({ itemsInCategory }) {
   return (
-    <>
+    <Paper elevation={1} sx={{ padding: '1rem' }}>
       <RowHeader title="Graph" caption="Graph details for last 10 recently updated" />
       <Box sx={{ position: 'relative', width: 'calc(100% - 1rem)' }}>
         <BarChart
           legendLabel="Name Vs Cost"
           data={
-            totalItems
+            itemsInCategory
               ?.filter((_, index) => index < 10)
               ?.map((v, index) => ({
                 label: v.name,
@@ -22,6 +23,6 @@ export default function ItemGraph({ totalItems }) {
           borderColor="rgba(75, 192, 192, 1)"
         />
       </Box>
-    </>
+    </Paper>
   );
 }
