--- frontend/src/features/MaintenancePlanDetails/MaintenancePlanItem/ItemHeader/ItemHeader.jsx
@@ -1,11 +1,0 @@
-import RowHeader from '../../../../common/RowHeader';
-import DetailsCard from '../../../../common/ItemCard/DetailsCard';
-
-export default function ItemHeader({ label, caption, item, image }) {
-  return (
-    <>
-      <RowHeader title={label} caption={caption} />
-      <DetailsCard selectedItem={item} selectedImage={image} />
-    </>
-  );
-}
